# src/main/resources/static/yaml/rule_modification_service-openapi.yaml
openapi: "3.1.0"
info:
  title: "Rule Modification Service API"
  description: "rule_modification_service API"
  version: "1.0.0"
servers:
  - url: "http://localhost:8080/api/"
paths:
  /translate:
    post:
      summary: "POST api/translate"
      operationId: "translate"
      parameters:
        - name: "disableDefaultModelInfoLoad"
          in: "query"
          required: false
          description: "Flag to disable default model info loading"
          schema:
            type: "boolean"
        - name: "verify"
          in: "query"
          required: false
          description: "Flag to enable verification"
          schema:
            type: "boolean"
        - name: "optimization"
          in: "query"
          required: false
          description: "Flag to enable optimization"
          schema:
            type: "boolean"
        - name: "annotations"
          in: "query"
          required: false
          description: "Flag to include annotations"
          schema:
            type: "boolean"
        - name: "locators"
          in: "query"
          required: false
          description: "Flag to include locators"
          schema:
            type: "boolean"
        - name: "resultTypes"
          in: "query"
          required: false
          description: "Flag to include result types"
          schema:
            type: "boolean"
        - name: "detailedErrors"
          in: "query"
          required: false
          description: "Flag to include detailed errors"
          schema:
            type: "boolean"
        - name: "errorLevel"
          in: "query"
          required: false
          description: "The severity level of errors"
          schema:
            type: "string"
            enum:
              - "Info"
              - "Warning"
              - "Error"
        - name: "disableListTraversal"
          in: "query"
          required: false
          description: "Flag to disable list traversal"
          schema:
            type: "boolean"
        - name: "disableListDemotion"
          in: "query"
          required: false
          description: "Flag to disable list demotion"
          schema:
            type: "boolean"
        - name: "disableListPromotion"
          in: "query"
          required: false
          description: "Flag to disable list promotion"
          schema:
            type: "boolean"
        - name: "enableIntervalDemotion"
          in: "query"
          required: false
          description: "Flag to enable interval demotion"
          schema:
            type: "boolean"
        - name: "enableIntervalPromotion"
          in: "query"
          required: false
          description: "Flag to enable interval promotion"
          schema:
            type: "boolean"
        - name: "disableMethodInvocation"
          in: "query"
          required: false
          description: "Flag to disable method invocation"
          schema:
            type: "boolean"
        - name: "requireFromKeyword"
          in: "query"
          required: false
          description: "Flag to require 'from' keyword in queries"
          schema:
            type: "boolean"
        - name: "strict"
          in: "query"
          required: false
          description: "Flag to enforce strict type checking"
          schema:
            type: "boolean"
        - name: "validateUnits"
          in: "query"
          required: false
          description: "Flag to validate units"
          schema:
            type: "boolean"
        - name: "signatures"
          in: "query"
          required: false
          description: "The signature level for library builder"
          schema:
            type: "string"
            enum:
              - "None"
              - "Differing"
              - "Overloads"
              - "All"
        - name: "compatibilityLevel"
          in: "query"
          required: false
          description: "The compatibility level for the translation"
          schema:
            type: "string"
            enum:
              - "1.3"
              - "1.4"
              - "1.5"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParamInjectionRequestBody"
              example: >
                {
                  "rule": {
                    "id": "LibraryOne",
                    "version": "1",
                    "content": "bGlicmFyeSBMaWJyYXJ5T25lIHZlcnNpb24gJzEnCgp1c2luZyBGSElSIHZlcnNpb24gJzQuMC4xJwoKaW5jbHVkZSBGSElSSGVscGVycyB2ZXJzaW9uICc0LjAuMScKCmNvbnRleHQgUGF0aWVudAoKZGVmaW5lICIkQWdlX2xpbWl0IjogMTAKCmRlZmluZSB4OiBBZ2VJblllYXJzKCkgPiAiJEFnZV9saW1pdCI="
                  }
                }
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
...
components:
  schemas:
    CqlRule:
      type: "object"
      properties:
        id:
          type: "string"
          description: CQL rule library id
        version:
          type: "string"
          description: CQL rule library version
        content:
          type: "string"
          description: Base64 encoded CQL rule
    ParamInjectionRequestBody:
      type: "object"
      properties:
        params:
          type: "object"
          additionalProperties:
            type: "object"
            properties:
              "type":
                type: "string"
              value:
                type: "any"

        rule:
          $ref: "#/components/schemas/CqlRule"
        libraries:
          type: "object"
          additionalProperties:
            type: "object"
            $ref: "#/components/schemas/CqlRule"